requirements.txt
pre-commit==3.7.1
sqlmodel==0.0.18
kink==0.8.0
pydantic-settings==2.2.1
playwright==1.44.0
typer==0.12.3
opencv-python==4.10.0.82
Flask[async]==3.0.3
pip==24.0
